Add files via upload
This is a program for determining the identity of partially degraded protein samples seen on mass spectra. Does not work well if degradation is too extreme, but excellent for if one is losing tags or small disordered regions on terminal ends of the protein. It is uploaded with an example sequence and mass to find.